revert ts to js
